Split DB, type enum
