+++ src/index.ts
@@ -4,7 +4,7 @@ import { api, bree as breeConfig, configDebug, socket } from "@/config";
 import bree from "@/scheduler";
 import socketServer from "@/socket";
 import app from "@/app";
-import { createConnection } from "@/services/db";
+import { createConnection } from "@/services/db/setup";
 import { getLogger } from "@/services/winston";
 
 const logger = getLogger("server-startup");
